Fix: Removed Logger and moved filter to stateless component
