+++ src/components/home.js
@@ -1,15 +1,16 @@
-import React, { useEffect, useState } from 'react';
+import React, { useEffect } from 'react';
 import { useDispatch, useSelector } from 'react-redux';
 import { NavLink } from 'react-router-dom';
 import { filterData, getData } from '../redux/home/home';
 import pointer from '../images/assets/pointer.svg';
+import Filter from './filter';
 
 const Home = () => {
   const data = useSelector((data) => data.detailDataReducer);
   const onLoad = useSelector((data) => data.loadingStateReducer);
+  const filterState = useSelector((data) => data.filterStateReducer);
   const dispatch = useDispatch();
   const mainArray = ['AAPL', 'FB', 'INTC', 'ORCL', 'NKE', 'PFE', 'NOK', 'TWTR', 'CSCO', 'BABA', 'ATVI', 'FOXA'];
-  const [filterState, setFilterState] = useState(0);
   useEffect(() => {
     dispatch(getData(mainArray));
   }, []);
@@ -17,12 +18,6 @@ const Home = () => {
   const handleClick = (props) => {
     dispatch(filterData(props));
   };
-
-  const handleChange = () => {
-    const filterDis = document.getElementById('filterDisplay');
-    setFilterState(filterDis.value);
-  };
-
   return (
     <div>
       <div className="Home_Header_1">
@@ -31,15 +26,7 @@ const Home = () => {
       </div>
       <div className="Home_Header_2">
         <h1>STATS BY COMPANY</h1>
-        <select id="filterDisplay" className="filter" onChange={() => { handleChange(); }}>
-          <option disabled selected value="0">Filter by Value</option>
-          <option value="0">All</option>
-          <option value="50">50+</option>
-          <option value="100">100+</option>
-          <option value="150">150+</option>
-          <option value="200">200+</option>
-          <option value="250">250+</option>
-        </select>
+        <Filter />
       </div>
       {onLoad ? (
         <div className="loading_state">
@@ -48,7 +35,7 @@ const Home = () => {
       ) : ' ' }
       <ul className="home_btns_cont">
         {data.map((e) => (
-          (e.Summary.Price >= parseInt(filterState, 10))
+          (e.Summary.Price >= parseInt(filterState.state, 10))
             ? (
               <li key={e.Summary.Name}>
                 <NavLink onClick={() => { handleClick(e); }} className="home_item_btn" type="button" key={e.Symbol} to="/details">
