+++ src/redux/configureStore.js
@@ -1,24 +1,24 @@
 import {
   createStore, compose, combineReducers, applyMiddleware,
 } from 'redux';
-import logger from 'redux-logger';
 import {
   fetchDataMiddleware,
   detailDataReducer,
   filteredDataReducer,
   loadingStateReducer,
   loadingChangeStateMiddleware,
+  filterStateReducer,
 } from './home/home';
 
 const reducer = combineReducers({
   detailDataReducer,
   filteredDataReducer,
   loadingStateReducer,
+  filterStateReducer,
 });
 const composedEnhancer = compose(
   applyMiddleware(fetchDataMiddleware),
   applyMiddleware(loadingChangeStateMiddleware),
-  applyMiddleware(logger),
 );
 
 const store = createStore(
